Fix not being able to order status section
Since the status title had extra, non-static information, it couldn't be
easily referenced in a custom section order. Strictly speaking, it was
nearly impossible due to the ANSI codes included to color properly.

Now all built-in state extensions support these accents. The accent is
independent of the title and only the title needs to be included when
supplying a custom order.

User defined sections do not support accents at this time.

Fixes #23